fix: EXPOSED-350 keepLoadedReferencesOutOfTransaction causes duplicate query when true (#2064)
Bug fix in version 0.45.0 introduced a side effect. The query for loading child
references would be executed twice if `keepLoadedReferencesOutOfTransaction = true`.

This happens because the select query was being sent to the database without
checking the cache first for loaded references.

Switched to a search method that goes through the cache first and made sure the
bug fix test also checked that each query was only run once.