fix: EXPOSED-384 CurrentTimestamp cannot be used with OffsetDateTimeColumnType
Since the 0.50.0 column type safety changes, CurrentTimestamp is no longer a class,
but an object, and it no longer returns type T, but Instant.

So its previous use as a default expression with a timezone compatible column is
no longer possible, as even changing to a singleton still required OffsetDateTime
as the return type.

This introduces a new object CurrentTimestampWithTimeZone for this specific use
to both modules exposed-java-time and exposed-kotlin-datetime.

Rather than repeating the same query builder logic, the 3 CurrentX objects now
extend from a base sealed class.