fix: EXPOSED-402 ClassCastException when eager loading with uuid().references()
Since version 0.50.0, using `Column.references()` invoked on a `UUIDColumnType`
that targets an `EntityIDColumnType<UUID>` causes a `ClassCastException` when
references are eager loaded, for the same type mismatch as detailed in the previous
issue.

This extends the previous fix to include Reference objects in `preloadRelations()`,
by forcing the query clause to use the underlying/wrapped type of `EntityIDColumnType`
if there is a type mismatch between `refColumn` and `refIds`.