fix: EXPOSED-400 ClassCastException when using `fetchBatchedResults` (#2113)
In the `fetchBatchedResults` function, the wrong `greater` and `less` methods were being invoked because `autoIncColumn`'s type is erased (`Column<*>`) and unknown at compile time. The issue is fixed by checking if `autoIncColumn.columnType` is an instance of `EntityIDColumnType` and casting accordingly to enable selecting the correct `greater` and `less` methods.