fix: EXPOSED-407 compositeMoney() nullability definition issues (#2137)
* fix: EXPOSED-407 compositeMoney() nullability definition issues

Defining a composite money column as nullable returns a different type, as
each component column is made nullable, but not the actual expression itself.

This fix adds an overload that returns the expected type and ensures that the
expression is also made to be nullable overall. Tests for using the component
columns individually have also been included.