fix: Allow column reference in default expressions for MySQL and MariaDB
MySQL and MariaDB are the only supported databases that allow referencing another column in a default expression, but for this to work, the referenced column must be one that occurs earlier in the table definition. This was not working because Exposed was sorting the columns by column name for the insert statement. Not sorting fixes this issue.