fix: EXPOSED-472 Alias IdTable fails with isNull and eq ops
The introduction of `CompositeIdTable`, and functions like `mapIdComparison` an>
`mapIdOperator`, led to a regression. When any `IdTable` is aliased and used wi>
the affected comparison operators, `eq` and `isNull` (and negative variants),
the following is thrown:
org.jetbrains.exposed.sql.Alias cannot be cast to org.jetbrains.exposed.dao.id.>

The aliased `idColumns` of the `Alias` delegate need to be used in these functi>
so these internal functions have been brought a level down, into the `Table` cl>
so they can be overridden in `Alias` as well.

`Alias.get()` operator has also been fixed to allow `CompositeIdTable` to be al>
and still access the delegate `id` column properly.