fix: EXPOSED-467 Decimal type precision and scale not checked by SchemaUtils
Any difference column type that includes a user-defined column size (precision)
or scale in the SQL type is not detected as a ColumnDiff when SchemaUtils functions
are used.

Column metadata for the fractional digits (scale) is now also read and use with
the existing column size to determine if an alter statement should be generated.