feat!: EXPOSED-320 Many-to-many relation with extra columns
An intermediate table is defined to link a many-to-many relation between 2 IdTables,
with references defined using via(). If this intermediate table is defined with
additional columns, these are not accessible through the linked entities.

This PR refactors the InnerTableLink logic to include additional columns in generated
SQL, which can be accessed as a regular field on one of the referencing entities.
It also allows the possibility to access the additional data as a new entity type,
which wraps the main child entity along with the additional fields. This is accomplished
with the introduction of InnerTableLinkEntity.