fix: EXPOSED-492 Eq/Neq op with partial composite key unwrapped value fails
Using eq()/neq() between a single component column of a CompositeIdTable, with type
EntityIDColumnType, and an unwrapped non-entityID value results to a cast exception.
The equality overload assumes that the value must be a CompositeID value, when it
is actually a value of one of the key's column, so the check for use of `mapIdComparison`
is restricted further as apropriate.