fix: EXPOSED-496 reference() idColumn equality check with referree's id is insufficient
isEntityIdentifier() relies currently on a column type to check whether that expression
is the id column of an idTable. This fails if the column being checked is a reference
entity id column to a referree table column that has the identical name. This happens
because column equality is solely based on table and name.

The function now relies on a column, so that it's originating table can be taken
more into account, thereby ensuring no false positives.