feat!: EXPOSED-497 Allow OFFSET without LIMIT in query
All dialects except SQLite and MySQL/MariaDB allow OFFSET clause to be used at the
end of a query without an associated LIMIT clause. This is currently not possible
to accomplish without creating a custom query or providing a very large number,
as a limit must be provided to get an offset.

This splits SizedIterable.limit() into 2 respective functions, introducing a breaking
change for the interface. It also creates a new FunctionProvider.queryLimitAndOffset()
to replace the original, so that a nullable limit can be passed.