docs: EXPOSED-515 How to identify composite key columns that use reference()
The documentation mentions that component columns of a composite key in
`CompositeIdTable` should be marked using `entityId()`. If this is done on columns
that are already `Column<EntityID<?>>`, it will lead to another nested wrap. This
can be avoided by marking the component column using `addIdColumn()` in the table
definition. An example of this using key columns that reference another table
has been included in both topics that explain how to define a `CompositeIdTable`.