fix: EXPOSED-513 DROP SEQUENCE fails when there is a dot in the sequence name (#2220)
When attempting to drop all available sequences in Oracle and H2 Oracle, DROP SEQUENCE fails when there is a dot in the sequence name because the name obtained from the database is not quoted. The error thrown for Oracle states that the sequence is not found in the database. The error thrown for H2 Oracle states that the schema is not found in the database.
In the overridden `sequences()` function in `OracleDialect` and `H2Dialect`, there is now a check to see if the sequence name has a dot, and if it does and it's not already quoted, quotes are added. This way, when the sequence's `identifier` is used to drop it in the sequence's drop statement, it would succeed in finding and dropping the sequence.