fix: EXPOSED-565 Subquery alias with id fails to use correct alias with eq
A fix for using mapIdComparison() with aliased table had been implemented, but
it did not fully cover the case when the right-hand side of the comparison involved
an unwrapped id column value. The latter was still using the delegate table associated
with the column type, instead of the invoking alias expression's alias table
identifier.