fix: EXPOSED-278 Invalid Oracle statement when adding a new column that is used in a primary key (#2259)
When adding a new column to an existing table in Oracle, if that new column is used in a primary key, then a separate statement is needed for the primary key otherwise there will be a syntax error. Previously, adding the column and the constraint was being done in one statement.