fix: EXPOSED-602 Column name that includes table name is aliased with upserts
Dialects that use MERGE for upserts automatically have their update values aliased.
Since version 0.54.0, update values that should be the same as insert values are
wrapped as a special expression. This means they are processed like any expression
and any potential inclusion of the table name is replaced with alias 'T'. This
causes invalid identifiers if the actual column to update uses an identifier
that includes the table name. Restricting the case for string replacement
fixes this.