docs: EXPOSED-670 Adjust YouTrack issue visibility and PR guidelines
Issues used to not have public visibility and this was made note of in docs, instead of
changing this. Issues can now be viewed without logging in and no documented links
should result in a 404 error.

PR guideline section was also updated to make clear that existing issues should
be used and add extra missing details.