docs: EXPOSED-207 Add link to SQLite ALTER TABLE restrictions in SchemaUtils Kdocs
Exposed generates ALTER TABLE ADD COLUMN or DROP COLUMN statements when needed by
certain SchemaUtils and MigrationUtils methods. For SQLite, there are many cases
when certain columns are not allowed to be used with these types of statements and
Exposed does not protect from them all. There is already a note in associated KDocs
that warns users of these statements failing, but it does not clearly state the
database restrictions. A link to the reference is now included.