fix: EXPOSED-696 [PostgreSQL] Drop of auto-increment sequence fails after column modified without dropping default (#2369)
DDL generated by MigrationUtils when a column is modified to lose its auto-increment
status was not dropping the column's default. This meant that attemtping to drop
the auto-increment status after would result in an exception because the sequence
was still being used by the altered column.

This ensures that the column is altered to drop its default if the existing database
column is auto-incrementing while the defined column isn't.