fix: EXPOSED-704 ClassCastException when referencing an eager-loaded backReferencedOn with keepLoadedReferencesOutOfTransaction = true
Both BackReference and OptionalBackReference classes use Referrers as their delegate
class when getValue() is invoked. This means that, when there is no transaction context,
the delegate correctly attempts to get a child value from the cache, but incorrectly
assumes it will be a SizedIterable (normal for Referrers).

This ensures that an attempt is made to store the child value as a LazySizedCollection
if it is not found to be a collection itself already.