refactor: EXPOSED-728 [SQLite] Remove ENABLE_UPDATE_DELETE_LIMIT metadata check from core function provider
SqliteDialect has a companion property that sends a low-level JDBC connection
query to check if LIMIT clause is allowed in either update or delete statements.
In preparation for R2DBC, this metadata query is removed from the core module and
replaced with appropriate DatabaseDialect and ExposedDatabaseMetadata functions that
can be called in appropriate blocking/non-blocking ways.

Tests have been adjusted accordingly.