fix: EXPOSED-737 Timestamp column broken for MySQL 8.0
In this [PR](https://github.com/JetBrains/Exposed/pull/2386), I modified the notNullValueToDB function to return the formatted String for MySQL (excluding MariaDB) because there was an issue with MySQL 5 where it implicitly converted the value to the session time zone when storing it and lead to an unexpected behaviour when the value is retrieved. This should have been done for MySQL 5 only and not for MySQL 8.0. This commit fixes that.