fix: Inserts execute without collecting & throwing expected exception
If an insert fails (constraint violation for example), it does not throw an exception
to be caught in certain dialects (SQL Server, Oracle, MariaDB). This is because those
dialects do not enter the logic branch that ensures the execution result is collected
from the flow (via mapRows()). After executeUpdate() is called, nothing else happens
to the result, so it's error is not collected.

This ensures that the appropriate branches now call a terminal operation to collect the flow.
It uses rowsUpdated() to do so for convenience.