test: Fix spring-boot-starter test block following bump to 3.5.0
Following spring-boot-starter bump to 3.5.0, all branches off main fail on TC
once they attempt to run tests in that module. This happens because of previously automatic
test dependency loading behavior change. Eventhough Exposed relies on JUnit 4,
one of the dependencies must use junit-jupiter, so the Gradle solution to add a
test runtime dependency on the platform launcher solves the issue.