revert: insert() and insertAndGetId() body parameter type
Prior to version 1.0.0, both these methods had a body parameter of function type,
which hadd an InsertStatement as the argument. This was changed to have an UpdateBuilder instead,
to accommodate the new statement builder. This means usage of any specific class properties
or methods in the lambda would no longer be possible. This may be unlikely but there
seemed no good reason for the change. The signatures have now been reverted.