fix: Provide typed exec(query) to avoid casting results in lamdba block (#2584)
* fix: Provide typed exec(query) to avoid casting results in lamdba

Pre-V1 it was possible to store a query (or any statement that returned results)
and then pass it to exec(). Within the latter's lambda block, the java.sql.ResultSet
was passed as the argument, so that it could be used directly to transform the results.

Post-V1, this is no longer possible as ResultApi is passed as the lambda argument.
So with JDBC, casting and unwrapping the ResultSet is first required.

This introduces typed execQuery() that does the casting and unwrapping under-the-hood.

* fix: Provide typed exec(query) to avoid casting results in lamdba

- Replace all references to specific beta versions