fix!: Replace SqlExpressionBuilder methods with top-level functions  (#2600)
* fix!: Replace SqlExpressionBuilder methods with top-level functions

Using SqlExpressionBuilder methods outside of designated scope blocks leads to
multiple issue imports on the object. There is also no clear pattern as to what
should become a method in this interface or not. The builder itself is also stateless
and serves no purpose beyond restricting context and reducing namespace pollution.

It has now been deprecated, along with Op.build() and Expression.build(), and its
implementations, in favor of top-level functions that can be used anywhere, even
with imports on demand.

* fix!: Replace SqlExpressionBuilder with top-level functions

- Update all samples & documentation with upated state