fix: EXPOSED-880 [PostgreSQL] Fix casting nullable JSON columns (#2616)
* EXPOSED-880: Fix casting nullable JSON columns

Exposed uses the first set of values in a batch statement when preparing
the statement. If one of those columns is a _nullable_ `json` or `jsonb`
column, and the SQL Dialect is Postgres, Exposed will neglect to cast
the placeholder to `json`/`jsonb`, and instead just render `?`.

There's no need for the `&& value != null` check here, because Postgres
will accept `null` values being cast as `json`/`jsonb`, if the column is
nullable.

* fix: EXPOSED-880 [PostgreSQL] Fix casting nullable JSON columns

- Add batch operation to nullable json/jsonb unit tests

---------

Co-authored-by: Chantal Loncle <82039410+bog-walk@users.noreply.github.com>